# modelcraft-backend/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ModelCraft REST API
  description: |
    REST API for tenant management (Auth & Organization) in the ModelCraft platform.

    Business domain APIs (Projects, Models, Clusters, Enums) are served exclusively
    via GraphQL at /org/{orgName}/design/graphql. This OpenAPI spec covers only identity and
    tenant-management endpoints.
  version: 3.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - BearerAuth: []
paths:
...
  /api/auth/refresh:
    post:
      operationId: refreshToken
      summary: Refresh tokens using httpOnly cookie (rotation)
      tags:
        - Auth
      security: []
      responses:
        '200':
          description: Token refresh successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Invalid, expired, or reused refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemError'
  /api/end-user/auth/login:
    post:
      operationId: endUserLogin
      summary: End-user login with username and password
      tags:
        - EndUserAuth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndUserLoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndUserLoginResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemError'
        '401':
          description: Invalid credentials or account disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndUserAuthError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemError'
  /api/end-user/auth/register:
    post:
      operationId: endUserRegister
      summary: End-user self-registration
      tags:
        - EndUserAuth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndUserRegisterRequest'
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndUserTokenResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemError'
        '401':
          description: Registration not allowed (cluster not configured)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndUserAuthError'
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndUserAuthError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemError'
  /api/end-user/auth/logout:
    post:
      operationId: endUserLogout
      summary: End-user logout (revoke refresh token via httpOnly cookie)
      tags:
        - EndUserAuth
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndUserLogoutRequest'
      responses:
        '204':
          description: Logout successful
  /api/end-user/auth/refresh:
    post:
      operationId: endUserRefreshToken
      summary: Refresh end-user access token using httpOnly refresh cookie
      tags:
        - EndUserAuth
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndUserRefreshRequest'
      responses:
        '200':
          description: Token refresh successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndUserTokenResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndUserAuthError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemError'
  /api/end-user/auth/select-project:
    post:
      operationId: endUserSelectProject
      summary: Select a project context for the end-user session
      tags:
        - EndUserAuth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndUserSelectProjectRequest'
      responses:
        '200':
          description: Project selected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndUserSelectProjectResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemError'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndUserAuthError'
        '403':
          description: No access to requested project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndUserAuthError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemError'
  /api/end-user/auth/me:
    get:
      operationId: endUserMe
      summary: Get current end-user profile (resolved from Bearer JWT)
      tags:
        - EndUserAuth
      security:
        - BearerAuth: []
      responses:
        '200':
          description: End-user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndUserMeResponse'
        '401':
          description: Missing or invalid Bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndUserAuthError'
        '403':
          description: Account disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndUserAuthError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemError'
  /api/user/memberships:
    get:
      summary: Get current user's organization memberships
      description: Returns a list of organizations the authenticated user is a member of, including role and join date information. Returns at most 100 memberships.
      operationId: getUserMemberships
      tags:
        - User
      responses:
        '200':
          description: List of user memberships
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMembershipsResponse'
        '401':
          description: Unauthorized - Invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    BaseResponse:
      type: object
      required:
        - requestId
      properties:
        requestId:
          type: string
          description: Unique request trace ID
          example: req_550e8400e29b41d4
    SystemError:
      type: object
      required:
        - requestId
        - error
      properties:
        requestId:
          type: string
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - SYSTEM_ERROR
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    UnauthorizedError:
      type: object
      required:
        - requestId
        - error
      properties:
        requestId:
          type: string
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - UNAUTHORIZED
            message:
              type: string
...
    AuthenticationError:
      type: object
      required:
        - requestId
        - error
      properties:
        requestId:
          type: string
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - AUTHENTICATION_FAILED
            message:
              type: string
...
    RefreshResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - accessToken
            - expiresIn
          properties:
            accessToken:
              type: string
              description: New short-lived JWT access token
            expiresIn:
              type: integer
              description: Access token TTL in seconds
    EndUserLoginRequest:
      type: object
      required:
        - orgName
        - username
        - password
      properties:
        orgName:
          type: string
          description: Organization slug
          example: my-org
        projectSlug:
          type: string
          description: Optional project slug for single-project login flow
          example: my-project
        username:
          type: string
          example: alice
        password:
          type: string
          example: s3cr3t
    EndUserRegisterRequest:
      type: object
      required:
        - orgName
        - projectSlug
        - username
        - password
      properties:
        orgName:
          type: string
          example: my-org
        projectSlug:
          type: string
          example: my-project
        username:
          type: string
          example: alice
        password:
          type: string
          example: s3cr3t
    EndUserLogoutRequest:
      type: object
      properties:
        orgName:
          type: string
          example: my-org
    EndUserRefreshRequest:
      type: object
      properties:
        orgName:
          type: string
          description: Required to locate the correct tenant DB for session lookup
          example: my-org
    EndUserSelectProjectRequest:
      type: object
      required:
        - orgName
        - projectSlug
      properties:
        orgName:
          type: string
          example: my-org
        projectSlug:
          type: string
          example: my-project
        refreshToken:
          type: string
          description: Refresh token (gateway reads from httpOnly cookie; direct callers may pass inline)
    EndUserTokenResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - userId
            - accessToken
            - expiresIn
          properties:
            userId:
              type: string
            accessToken:
              type: string
              description: Short-lived JWT access token
            expiresIn:
              type: integer
              description: Access token TTL in seconds
            expiresAt:
              type: string
              format: date-time
            projects:
              type: array
              items:
                $ref: '#/components/schemas/EndUserAccessibleProject'
              description: Accessible projects (returned on login when projectSlug not specified)
    EndUserLoginResponse:
      allOf:
        - $ref: '#/components/schemas/EndUserTokenResponse'
    EndUserSelectProjectResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - selectedProject
          properties:
            selectedProject:
              type: string
              description: The selected project slug
    EndUserMeResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - endUser
          properties:
            endUser:
              $ref: '#/components/schemas/EndUserProfile'
    EndUserProfile:
      type: object
      required:
        - id
        - username
        - isForbidden
      properties:
        id:
          type: string
        username:
          type: string
        isForbidden:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EndUserAccessibleProject:
      type: object
      required:
        - slug
        - title
      properties:
        slug:
          type: string
        title:
          type: string
    EndUserAuthError:
      type: object
      required:
        - requestId
        - error
      properties:
        requestId:
          type: string
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code, e.g. AUTHENTICATION_FAILED, OPERATION_DENIED.END_USER
            message:
              type: string
    GetMembershipsResponse:
      type: object
      required:
        - requestId
        - memberships
      properties:
        requestId:
          type: string
          description: Unique request identifier
          example: req-12345678-1234-1234-1234-123456789012
        memberships:
          type: array
          items:
            $ref: '#/components/schemas/MembershipInfo'
          description: List of organization memberships
    MembershipInfo:
      type: object
      required:
        - orgId
        - orgName
        - displayName
        - role
        - joinedAt
      properties:
        orgId:
          type: string
          description: Organization UUID
          example: 550e8400-e29b-41d4-a716-446655440000
        orgName:
          type: string
          description: Organization slug (unique identifier)
          example: modelcraft
        displayName:
          type: string
          description: Organization display name
          example: ModelCraft
        role:
          type: string
          description: User's role in the organization
          example: owner
        joinedAt:
          type: integer
          format: int64
          description: Join timestamp in milliseconds (Unix epoch)
          example: 1708435200000
...